+++ doc/changes.rst
@@ -5,7 +5,13 @@ desitarget Change Log
 0.26.0 (unreleased)
 -------------------
 
-* No changes yet
+* More updates to target classes, mainly for SV [`PR #418`_]. Includes:
+    * First full implementations of `QSO`, `LRG`, `ELG`, and `STD` for SV.
+    * Update and refactor of `MWS` and `BGS` classes for the main survey.
+    * Change name of main survey `MWS_MAIN` class to `MWS_BROAD`.
+    * Augment QA code to handle SV sub-classes such as `ELG_FDR_FAINT`.
+
+.. _`PR #418`: https://github.com/desihub/desitarget/pull/418
 
 0.25.0 (2018-11-07)
 -------------------